Posts api endpoints to frontend for home page

See merge request I454903/s3-g3-feedbackapp!8

+++ BE/src/main/java/com/g3/feedbackApp/DataSources/Interfaces/IDataSourceReviewer.java
@@ -5,9 +5,9 @@ import com.g3.feedbackApp.Models.ReviewerModel;
 import java.util.List;
 
 public interface IDataSourceReviewer {
-    public ReviewerModel getReviewerById(int id);
+    public ReviewerModel getReviewerById(Long id);
     public List<ReviewerModel> getReviewers();
-    public boolean deleteReviewerById(int id);
+    public boolean deleteReviewerById(Long id);
     public boolean addReviewer(ReviewerModel reviewer);
     public boolean updateReviewer(ReviewerModel reviewer);
 }
